requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
huggingface-hub==0.19.4
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
streamlit==1.28.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
reportlab==4.0.4
python-multipart==0.0.6
